refactor: move shared content_libraries data classes into separate file
Code borrowed from cee5722fd73607708a2dce7d70598d57b2fe5bb
